lexbridge-ai-main/data_module/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data_module"
version = "0.1.0"
description = "Modular QA data ingestion, storage, and retrieval pipeline for RAG and knowledge graphs"
readme = "README.md"
requires-python = ">=3.10"
license = "CC-BY-SA-4.0"
dependencies = [
    # Core
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
    "tqdm>=4.66",
    # Data processing
    "pyarrow>=15.0",
    "pandas>=2.0",
    "duckdb>=0.10",
    "lancedb>=0.6",
    # Text processing
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
    "spacy>=3.7",
    "langdetect>=1.0",
    # Embeddings
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    # Graph
    "networkx>=3.2",
    # Hugging Face
    "datasets>=2.18",
    "huggingface-hub>=0.22",
    # HTTP / downloads
    "httpx>=0.27",
    "aiofiles>=23.2",
    "py7zr>=0.21",
    # BM25 / hybrid search
    "rank-bm25>=0.2",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.18"]
openai = ["openai>=1.23"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "black>=24.0", "ruff>=0.4", "mypy>=1.9"]

[project.scripts]
data-download = "data_module.scripts.download_sources:app"
data-pipeline = "data_module.scripts.run_pipeline:app"
data-index    = "data_module.scripts.build_index:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_module*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
